src/main/java/com/life/full/danger/sell/category/service/impl/CategoryServiceImpl.java
package com.life.full.danger.sell.category.service.impl;

import com.life.full.danger.sell.category.dao.ProductCategoryDao;
import com.life.full.danger.sell.category.model.ProductCategory;
import com.life.full.danger.sell.category.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * 类目
 * @author LiHongHui
 * @date 2018/9/14 9:56
 * @description
 */
@Service
public class CategoryServiceImpl implements CategoryService {
    @Autowired
    private ProductCategoryDao productCategoryDao;
    @Override
    public ProductCategory findOne(Integer categoryId) {
        return productCategoryDao.findById(categoryId).get();
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryDao.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return productCategoryDao.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryDao.save(productCategory);
    }
}
